+++ src/app/layout.tsx
@@ -23,6 +23,7 @@ import SuppressExtensionErrors from "@/components/SuppressExtensionErrors";
 import { TrackContextMenu } from "@/components/TrackContextMenu";
 import { UIWrapper } from "@/components/UIWrapper";
 import { AudioPlayerProvider } from "@/contexts/AudioPlayerContext";
+import { AuthModalProvider } from "@/contexts/AuthModalContext";
 import { MenuProvider } from "@/contexts/MenuContext";
 import { PlaylistContextMenuProvider } from "@/contexts/PlaylistContextMenuContext";
 import { ThemeProvider } from "@/contexts/ThemeContext";
@@ -118,53 +119,55 @@ export default function RootLayout({
           <SessionProvider>
             <TRPCReactProvider>
               <ThemeProvider>
-                <ElectronChromeSync />
-                <ToastProvider>
-                  <AudioPlayerProvider>
-                    {}
-                    <DynamicTitle />
-                    <MenuProvider>
-                      <TrackContextMenuProvider>
-                        <PlaylistContextMenuProvider>
-                          {}
-                          <UIWrapper>
+                <AuthModalProvider>
+                  <ElectronChromeSync />
+                  <ToastProvider>
+                    <AudioPlayerProvider>
+                      {}
+                      <DynamicTitle />
+                      <MenuProvider>
+                        <TrackContextMenuProvider>
+                          <PlaylistContextMenuProvider>
                             {}
-                            <DesktopShell>
-                              <div suppressHydrationWarning>
-                                <Suspense fallback={null}>
-                                  <Header />
-                                </Suspense>
-                              </div>
+                            <UIWrapper>
                               {}
-                              <Suspense fallback={null}>
-                                <MobileHeader />
-                              </Suspense>
-                              {}
-                              <HamburgerMenu />
-                              {}
-                              <MobileContentWrapper>
-                                {}
-                                <div className="pt-16 pb-36 md:pt-0 md:pb-24">
-                                  {children}
+                              <DesktopShell>
+                                <div suppressHydrationWarning>
+                                  <Suspense fallback={null}>
+                                    <Header />
+                                  </Suspense>
                                 </div>
-                              </MobileContentWrapper>
-                            </DesktopShell>
-                          </UIWrapper>
-                          {}
-                          <PersistentPlayer />
-                          {}
-                          <Suspense fallback={null}>
-                            <MobileFooterWrapper />
-                          </Suspense>
-                          {}
-                          <TrackContextMenu />
-                          {}
-                          <PlaylistContextMenu />
-                        </PlaylistContextMenuProvider>
-                      </TrackContextMenuProvider>
-                    </MenuProvider>
-                  </AudioPlayerProvider>
-                </ToastProvider>
+                                {}
+                                <Suspense fallback={null}>
+                                  <MobileHeader />
+                                </Suspense>
+                                {}
+                                <HamburgerMenu />
+                                {}
+                                <MobileContentWrapper>
+                                  {}
+                                  <div className="pt-16 pb-36 md:pt-0 md:pb-24">
+                                    {children}
+                                  </div>
+                                </MobileContentWrapper>
+                              </DesktopShell>
+                            </UIWrapper>
+                            {}
+                            <PersistentPlayer />
+                            {}
+                            <Suspense fallback={null}>
+                              <MobileFooterWrapper />
+                            </Suspense>
+                            {}
+                            <TrackContextMenu />
+                            {}
+                            <PlaylistContextMenu />
+                          </PlaylistContextMenuProvider>
+                        </TrackContextMenuProvider>
+                      </MenuProvider>
+                    </AudioPlayerProvider>
+                  </ToastProvider>
+                </AuthModalProvider>
               </ThemeProvider>
             </TRPCReactProvider>
           </SessionProvider>